requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
dateparser==0.7.6
dnspython==1.16.0
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.0
numpy==1.19.0
Pillow==7.2.0
PyJWT==1.7.1
pymongo==3.10.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
regex==2020.7.14
requests==2.24.0
six==1.15.0
twilio==6.44.0
tzlocal==2.1
urllib3==1.25.9
Werkzeug==1.0.1
wordcloud==1.7.0
